+++ src/commands/modmail/close.ts
@@ -3,7 +3,7 @@ import BasicEmbed from "../../utils/BasicEmbed";
 import Modmail from "../../models/Modmail";
 import { waitingEmoji } from "../../Bot";
 import { ThingGetter } from "../../utils/TinyUtils";
-import Database from "../../utils/cache/database";
+import Database from "../../utils/data/database";
 import log from "fancy-log";
 import { CommandOptions, SlashCommandProps } from "commandkit";
