+++ src/commands/roles/dontatmerole.ts
@@ -6,7 +6,7 @@ import {
   RepliableInteraction,
 } from "discord.js";
 import log from "fancy-log";
-import Database from "../../utils/cache/database";
+import Database from "../../utils/data/database";
 import DontAtMeRole from "../../models/DontAtMeRole";
 import { waitingEmoji } from "../../Bot";
 import BasicEmbed from "../../utils/BasicEmbed";
