+++ src/events/interactionCreate/tagautocomplete.ts
@@ -3,7 +3,7 @@ import log from "fancy-log";
 import TagSchema from "../../models/TagSchema";
 import { upperCaseFirstLetter, getTagKey, getTagName, debugMsg } from "../../utils/TinyUtils";
 import { redisClient } from "../../Bot";
-import Database from "../../utils/cache/database";
+import Database from "../../utils/data/database";
 const COMMAND_NAME = "tag";
 
 export default async (interaction: AutocompleteInteraction, client: Client<true>) => {
